requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
pydantic-settings
pytest
pyyaml
httpx
pytest-asyncio
pytest-playwright
playwright
